Roster: use an array instead of a list
The current Elm array implementation is buggy when it comes to
appending and slicing [1], so we had to use a list which is less performant
in our use case, but it works correctly.

Skinney released a new array implementation for Elm [2] with all the bugs
fixed, so we can rely on it instead of using a list as an implementation
detail.

[1] https://github.com/elm-lang/core/issues/649
[2] https://github.com/Skinney/elm-array-exploration